refactor: dto 클래스의 패키지 이동 및 dto 패키지 삭제
