+++ src/main/java/chess/dao/PieceDto.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.dao;
 
 public record PieceDto(Long gameId, String pieceAppearance, String positionColumn, String positionRow) {
 }
